// frontend/src/core/types.ts
export interface ErrorResponse {
    message: string;
}
// ...
export interface User {
    id: number;
    created_at: string;
    last_seen: string;
    online: boolean;
    username: string;
    admin?: boolean;
    bio?: string;
}
// ...
export interface LoginRequest {
    username: string;
    password: string;
}
// ...
export interface LoginResponse {
    user: User;
    token: string;
}